feat: Add PUT route for updating a post
This commit adds a new PUT route in the post API to handle updating a post. The route is protected and can only be accessed by authenticated users with admin privileges. The route expects a JSON payload containing the updated post data, and it updates the corresponding post in the database. If the post is not found, an error is thrown. This new route allows for editing existing posts in the application.

+++ src/app/api/v1/post/[postId]/route.ts
@@ -31,12 +31,64 @@ export async function GET(
   }
 }
 
+export async function PUT(
+  request: NextRequest,
+  { params }: { params: { postId: string } }
+) {
+  const token = await getToken({
+    req: request,
+    secret: process.env.NEXTAUTH_SECRET,
+  });
+
+  if (token === null || token?.role !== "admin") {
+    return Response.json(
+      { success: false, msg: "Unauthorized" },
+      { status: 401 }
+    );
+  }
+
+  try {
+    await dbConnect();
+    const data = await request.json();
+    const postId = params.postId;
+    const post = await Post.findByIdAndUpdate(
+      {
+        _id: postId,
+      },
+      data
+    );
+    if (!post) {
+      throw new Error("Post not found");
+    }
+    return Response.json({ success: true });
+  } catch (error: any) {
+    const myBlob = {
+      success: false,
+      msg: error.message,
+    };
+    const myOptions = { status: 400 };
+    return Response.json(myBlob, myOptions);
+  }
+}
+
 export async function DELETE(
   request: NextRequest,
   { params }: { params: { postId: string } }
 ) {
-  await dbConnect();
+  const token = await getToken({
+    req: request,
+    secret: process.env.NEXTAUTH_SECRET,
+  });
+
+  if (token === null || token?.role !== "admin") {
+    return Response.json(
+      { success: false, msg: "Unauthorized" },
+      { status: 401 }
+    );
+  }
+
   try {
+    await dbConnect();
     const postId = params.postId;
     const post = await Post.findByIdAndDelete(postId);
     if (!post) {
